Make template visibility toggle consistent with overlays
This is a good first step to fixing this as it rules out key the key
itself as the source of the bug. I do hate this code, but using the same
bad code where it applies should hopefully make refactoring it all later
simpler.

Given that, this seems to be caused by some race condition to update the
url.