Fix template dragging
Ever since template dragging has existed in pxls, it has suffered the
bug of dragging too fast leaving the template behind. This is because
the cursor move events exceed the bounds of the element before it can be
moved and doesn't catch the next event.

This fixes this by binding the move listener to the whole screen.